analysis-service/requirements.txt
ast-tools==0.1.8
astor==0.8.1
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
Flask==3.1.1
flask-cors==6.0.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
libcst==1.8.1
lizard==1.17.31
mando==0.7.1
MarkupSafe==3.0.2
pathspec==0.12.1
Pygments==2.19.1
python-dotenv==1.1.0
PyYAML==6.0.2
radon==6.0.1
requests==2.32.4
six==1.17.0
SQLAlchemy==2.0.41
tree-sitter==0.24.0
tree-sitter-java==0.23.5
tree-sitter-javascript==0.23.1
tree-sitter-python==0.23.6
typing_extensions==4.14.0
urllib3==2.4.0
Werkzeug==3.1.3
